requirements.txt
streamlit
pandas
notebook
numpy
scipy
matplotlib
seaborn
ensure==1.0.2
python-box==6.0.2
pyYAML
types-pyYAML
joblib
tqdm
plotly
scikit-learn
scikit-surprise
-e .